requirements.txt
streamlit
docx2txt
PyPDF2
pdfplumber
python-docx
groq
